ending it all
Former-commit-id: 13e5d26acd234a11b462f3f94f5d20355570ad09 [formerly 13765ad6a4e7fa6065747f484e4ebab80111c17e] [formerly e847b47b0116d96c9c451211cb38d3dcc890ce19 [formerly 81cc50ad1d472cb4e4bd00063e190e5af216655c]]
Former-commit-id: b246466f8e5343e43a6fd2bbe883043faa3ac1bf
Former-commit-id: 604a8ff98d445d0bf44a07c410d836d110fe4f58

+++ contracts/lending/IVaultController.sol
@@ -1,26 +1,28 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
-
 /// @title VaultController Interface
 /// @notice Interface for interacting with the VaultController
 interface IVaultController {
   // view functions
-  function _totalBaseLiability() external view returns(uint256);
-  function _lastInterestTime() external view returns(uint256);
-  function _protocolFee() external view returns(uint256);
+  function _totalBaseLiability() external view returns (uint256);
+
+  function _lastInterestTime() external view returns (uint256);
+
+  function _protocolFee() external view returns (uint256);
+
   function InterestFactor() external view returns (uint256);
 
   function ProtocolFee() external view returns (uint256);
 
-  function VaultAddress(uint256 id) external view returns (address);
+  function VaultAddress(uint96 id) external view returns (address);
 
-  function AccountLiability(uint256 id) external view returns (uint256);
+  function AccountLiability(uint96 id) external view returns (uint256);
 
-  function AccountBorrowingPower(uint256 id) external view returns (uint256);
+  function AccountBorrowingPower(uint96 id) external view returns (uint256);
 
   function TokensToLiquidate(
-    uint256 id,
+    uint96 id,
     address token,
     uint256 num
   ) external view returns (uint256);
@@ -30,19 +32,19 @@ interface IVaultController {
 
   function mintVault() external returns (address);
 
-  function checkAccount(uint256 id) external view returns (bool);
+  function checkAccount(uint96 id) external view returns (bool);
 
   function liquidate_account(
-    uint256 id,
+    uint96 id,
     address asset_address,
     uint256 tokenAmount
   ) external returns (uint256);
 
-  function borrowUsdi(uint256 id, uint256 amount) external;
+  function borrowUsdi(uint96 id, uint256 amount) external;
 
-  function repayUSDi(uint256 id, uint256 amount) external;
+  function repayUSDi(uint96 id, uint256 amount) external;
 
-  function repayAllUSDi(uint256 id) external;
+  function repayAllUSDi(uint96 id) external;
 
   function calculateInterest() external returns (uint256);
